moved graceful shutdown to main
