+++ main.go
@@ -1,6 +1,11 @@
 package main
 
 import (
+	"context"
+	"os"
+	"os/signal"
+	"time"
+
 	"github.com/kosatnkn/catalyst/app/config"
 	"github.com/kosatnkn/catalyst/app/container"
 	"github.com/kosatnkn/catalyst/app/http/server"
@@ -19,5 +24,26 @@ func main() {
 	ctr := container.Resolve(cfg)
 
 	// start the server to handle requests
-	server.Run(cfg.AppConfig, ctr)
+	srv := server.Run(cfg.AppConfig, ctr)
+
+	// enable graceful shutdown
+	c := make(chan os.Signal, 1)
+
+	// accept graceful shutdowns when quit via SIGINT (Ctrl+C)
+	// SIGKILL, SIGQUIT or SIGTERM (Ctrl+/) will not be caught
+	signal.Notify(c, os.Interrupt)
+
+	// Block until we receive our signal
+	<-c
+
+	// Create a deadline to wait for
+	var wait time.Duration
+
+	ctx, cancel := context.WithTimeout(context.Background(), wait)
+	defer cancel()
+
+	// gracefully release all additional resources
+	server.Stop(ctx, srv, ctr)
+
+	os.Exit(0)
 }
